Added Extensions.gwt.xml to playn extensions library. Added moving chimneys. Added auto clear entities as soon as they leave the screen. Added tripleplay to dependencies. Added gwt configuration to generate the webapp. Added some filter methods and generic operations to world.*
*This should be carefully reviewed as this methods may be redundant and
non performant when compared with the update lifecycle. It should be decided
if all the transformations to the world must be done in the lifecycle or if
it should be possible to perform these transformations outside of the lifecycle,
namely in public methods in the world that can apply transformations to
entities previously filtered.